Check for external modules in container types
When looking for external modules to import into the file we were
only looking at *parser.ClassTypes and not looking to see if there
is a vector or map with an included type.

This adds a recursive lookup for vector item items, and map value types.
We are currently ignoring map key types as we only allow pointers to
classes and so key values would be pretty worthless.

Since jute will only really be used by zookeeper i think this is a fine
trade off as it doesn't use maps at all.