Make strings option by making them pointers
In Java strings can be nil and Jute uses this functionality for optional
strings.  To replicate this we can make all strings pointers.  It makes
the generate code a bit less friendly to use, but for the most part the
generate structs shouldn't be exposed to the end user so we can hide
this.